# method_pipeline/generated/chatdev/minimax-m2.7/generic/QA2-1207/sdk_chatdev_20260712091919_20260712091919/code_workspace/oms_backend/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management System API
  description: |
    A production-grade Order Management System backend implementing the complete order workflow.
    
    ## Workflow Steps
    1. Customer places order
    2. Order Staff reviews & accepts
    3. Accountant creates invoice
    4. Customer pays invoice
    5. Accountant verifies payment
    6. Order Staff ships paid order
    7. Order Staff closes completed order
    
    ## NFR Support
    - NFR 2.1 Graceful Degradation: Feature flags and circuit breakers
    - NFR 2.2 Fault Detection: Health checks and automatic recovery
    - NFR 2.3 State Preservation: WAL mode, snapshots, and idempotency keys
  version: 1.0.0
  contact:
    name: OMS Backend Team
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /api/v1/customers:
    get:
      tags: [customers]
      summary: List customers
      parameters:
        - name: skip
          in: query
          schema: {type: integer, default: 0}
        - name: limit
          in: query
          schema: {type: integer, default: 100}
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /api/v1/customers/{customer_id}:
    get:
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: customer_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
    patch:
      tags: [customers]
      summary: Update customer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
    delete:
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '204':
          description: Customer deleted
...
  /api/v1/orders:
    get:
      tags: [orders]
      summary: List orders
      parameters:
        - name: skip
          in: query
          schema: {type: integer, default: 0}
        - name: limit
          in: query
          schema: {type: integer, default: 100}
        - name: status
          in: query
          schema: {type: string}
        - name: customer_id
          in: query
          schema: {type: string}
      responses:
        '200':
          description: List of orders
    post:
      tags: [orders]
      summary: Place order (Step 1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order placed
  /api/v1/orders/{order_id}:
    get:
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Order found
  /api/v1/orders/{order_id}/accept:
    patch:
      tags: [orders]
      summary: Accept order (Step 2)
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Order accepted
  /api/v1/orders/{order_id}/ship:
    patch:
      tags: [orders]
      summary: Ship order (Step 6)
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
        - name: tracking_number
          in: query
          schema: {type: string}
      responses:
        '200':
          description: Order shipped
  /api/v1/orders/{order_id}/close:
    patch:
      tags: [orders]
      summary: Close order (Step 7)
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Order closed
  /api/v1/invoices:
    post:
      tags: [invoices]
      summary: Create invoice (Step 3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Invoice created
  /api/v1/invoices/{invoice_id}:
    get:
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Invoice found
...
components:
  schemas:
    Customer:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        address: {$ref: '#/components/schemas/Address'}
        role: {type: string}
        created_at: {type: string}
        updated_at: {type: string}
    Address:
      type: object
      properties:
        street: {type: string}
        city: {type: string}
        state: {type: string}
        postal_code: {type: string}
        country: {type: string}
    CreateCustomerRequest:
      type: object
      required: [name, email]
      properties:
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        address: {$ref: '#/components/schemas/Address'}
    UpdateCustomerRequest:
      type: object
      properties:
        name: {type: string}
        email: {type: string}
        phone: {type: string}
...
    PlaceOrderRequest:
      type: object
      required: [customer_id, line_items, shipping_address]
      properties:
        customer_id: {type: string}
        line_items: {type: array}
        shipping_address: {$ref: '#/components/schemas/Address'}
        notes: {type: string}
        idempotency_key: {type: string}
...
    CreateInvoiceRequest:
      type: object
      required: [order_id, customer_id, billing_address]
      properties:
        order_id: {type: string}
        customer_id: {type: string}
        billing_address: {$ref: '#/components/schemas/Address'}
        due_date_days: {type: integer, default: 30}
...